% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoDisc.R
\name{autoDisc}
\alias{autoDisc}
\title{autoDisc}
\usage{
autoDisc(
  data,
  y,
  train_subset = c(rep(1, nrow(data))),
  exclude = NULL,
  include = NULL,
  bs_loc = "equal",
  na_loc = "equal",
  min_r = 0.03,
  cp = 0,
  maxdepth = 3,
  file_name = "disc",
  deploy_r = FALSE,
  deploy_sql = FALSE,
  deploy_python = FALSE,
  pct_na_allowed = 0.9,
  quantil_variability_exclude = 0.015,
  trim = 0.01,
  return_data = TRUE,
  subsample_rows = 1,
  seed = NULL,
  debug = FALSE
)
}
\arguments{
\item{data}{R data frame object.}

\item{y}{Name of you target column.}

\item{train_subset}{Vector of 0 and 1, with the same len of data.}

\item{exclude}{If you do not want to pass the name of the Data Frame variables to be analyzed by the algorithm, you also have to pass the list of data frame variables that are not analyzed, for example the key variables, and everything else will be analyzed.}

\item{include}{List with name of variables to be analyzed by the algorithm.}

\item{bs_loc}{Where put baseline: "euqal", "min" or "max".}

\item{na_loc}{Where put NAs: "euqal", "min" or "max".}

\item{min_r}{min rep of the tree leaves.}

\item{cp}{cp of the tree.}

\item{maxdepth}{maxdepth of the tree.}

\item{file_name}{File name with the deploy.}

\item{deploy_r}{TRUE or FALSE.}

\item{deploy_sql}{TRUE or FALSE.}

\item{deploy_python}{TRUE or FALSE.}

\item{pct_na_allowed}{Pct of NA allowed in each covariable.}

\item{quantil_variability_exclude}{Pct of qunatil, to remove vars that not present variability before discretize.}

\item{trim}{Pct of trim vars, to remove outliers before discretize.}

\item{return_data}{A boolean indicate if function will return data frame.}

\item{subsample_rows}{Proportion of your data that function will run.}

\item{seed}{Seed to subsample rows.}

\item{debug}{debug.}
}
\value{
A data_frame with the vars DC.
}
\description{
This function auto disc your vars for regression.
}
\examples{
\dontrun{
data_return = autoDisc(data = your_df
                      ,y = "y"
                      ,train_subset = your_df$Flag_de
                      ,include = include
                      ,deploy_r = TRUE
                      ,deploy_sql = TRUE)
}
}
