% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGroup.R
\name{autoGroup}
\alias{autoGroup}
\title{autoGroup}
\usage{
autoGroup(
  data,
  y,
  train_subset = c(rep(1, nrow(data))),
  exclude = NULL,
  include = NULL,
  bs_loc = "equal",
  na_replace = "",
  min_r = 0.03,
  cp = 0,
  maxdepth = 3,
  file_name = "group",
  deploy_r = FALSE,
  deploy_sql = FALSE,
  pct_na_allowed = 0.95,
  perc_max_cat = 0.9,
  return_data = TRUE,
  warn = NULL,
  seed = NULL,
  subsample_rows = 1,
  debug = FALSE
)
}
\arguments{
\item{data}{R data frame object.}

\item{y}{Name of you target column.}

\item{train_subset}{Vector of 0 and 1, with the same len of data.}

\item{exclude}{If you do not want to pass the name of the Data Frame variables to be analyzed by the algorithm, you also have to pass the list of data frame variables that are not analyzed, for example the key variables, and everything else will be analyzed.}

\item{include}{List with name of variables to be analyzed by the algorithm.}

\item{bs_loc}{Where put baseline: "euqal", "min" or "max".}

\item{na_replace}{String to replace NAs, its recomended do this, to NA be included in a band. If na_replace = NA, NAs will be include in a band with min diff of avg y.}

\item{min_r}{min rep of the tree leaves.}

\item{cp}{cp of the tree.}

\item{maxdepth}{maxdepth of the tree.}

\item{file_name}{File name with the deploy.}

\item{deploy_r}{TRUE or FALSE.}

\item{deploy_sql}{TRUE or FALSE.}

\item{pct_na_allowed}{Pct of NA allowed in each covariable.}

\item{perc_max_cat}{Perc of nrow data that perc of number of categories of each var are allowed.}

\item{return_data}{A boolean indicate if function will return data frame.}

\item{warn}{Indicate if function will show warnings.}

\item{seed}{Seed to subsample rows.}

\item{subsample_rows}{Proportion of your data that function will run.}

\item{debug}{debug.}
}
\value{
A data_frame with the vars DC.
}
\description{
This function auto group your vars.
}
\examples{
\dontrun{
autoGroup(data = your_df
         ,y = "y"
         ,train_subset = your_df$Flag_dev
         ,include = include
         ,deploy_r = T
         ,deploy_sql = TRUE)
}
}
