% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_p.R
\name{dp_p}
\alias{dp_p}
\title{dp_p}
\usage{
dp_p(var_name, summary, mean.na, data_name = "data", file)
}
\arguments{
\item{var_name}{var name.}

\item{summary}{summary.}

\item{mean.na}{mean of NAs.}

\item{data_name}{data name.}

\item{file}{txt  to deploy.}
}
\description{
This function deploy autoDisc in python.
}
