% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_s.R
\name{dp_s}
\alias{dp_s}
\title{dp_s}
\usage{
dp_s(var_name, summary, data_name = "data", mean.na, file)
}
\arguments{
\item{var_name}{var name.}

\item{summary}{summary.}

\item{data_name}{data name.}

\item{mean.na}{mean of NAs.}

\item{file}{txt  to deploy.}
}
\description{
This function deploy autoDisc in SQL.
}
