% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpg_r.R
\name{dpg_r}
\alias{dpg_r}
\title{dpg_s}
\usage{
dpg_r(var_name, summary, mean.na, data_name = "data", file)
}
\arguments{
\item{var_name}{var name.}

\item{summary}{summary.}

\item{mean.na}{mean of NAs.}

\item{data_name}{data name.}

\item{file}{txt  to deploy.}
}
\description{
This function deploy autoDisc in r.
}
